move tests from tests_application/{test_payment, test_application_references, test_adult_in_home, test_login_redirect_helper, test_personal_details} to preferred structure. Refactor some tests
