+++ application/views/payment.py
@@ -59,7 +59,9 @@ def payment_confirmation(request):
     return render(request, template, variables)
 
 
+# TODO: sends email as a side-effect
 def get_template(crc, app_id, application, cost, cr_type):
+
     lived_abroad = crc.lived_abroad
     capita = crc.capita
 
@@ -83,16 +85,18 @@ def get_template(crc, app_id, application, cost, cr_type):
                        'firstName': first_name,
                        'cost': cost}
 
-    log.debug(
-        'Attempting to send a payment confirmation email with early_years_register: {0}; cr_type: {1}; cost: {2}; capita: {3}; cautions_convictions: {4}; lived_abroad: {5}'.format(
-            early_years_register, cr_type, cost, capita, certificate_information_boolean, lived_abroad))
+    log.debug(('Attempting to send a payment confirmation email with early_years_register: {0}; cr_type: {1}; '
+               'cost: {2}; capita: {3}; cautions_convictions: {4}; lived_abroad: {5}').format(
+                early_years_register, cr_type, cost, capita, certificate_information_boolean, lived_abroad))
+
     if (capita and not certificate_information_boolean) and lived_abroad:
         if early_years_register:
             log.debug('Attempting send of email "Confirmation with HDB - lived abroad only email"')
         else:
             log.debug('Attempting send of email "Confirmation - lived abroad only"')
 
-        email_template = '36720ba3-165e-40cd-a6d2-320daa9d6e4a' if early_years_register else 'f5a2998c-7322-4e32-8a85-72741bfec4a5'
+        email_template = '36720ba3-165e-40cd-a6d2-320daa9d6e4a' if early_years_register \
+                         else 'f5a2998c-7322-4e32-8a85-72741bfec4a5'
         view_template = 'payment-confirmation-lived-abroad.html'
 
     elif ((capita and certificate_information_boolean) or on_update) and lived_abroad:
@@ -101,7 +105,8 @@ def get_template(crc, app_id, application, cost, cr_type):
         else:
             log.debug('Attempting send of email "Confirmation - DBS & lived abroad"')
 
-        email_template = 'c82b8ffd-f67c-4019-a724-d57ab559f08e' if early_years_register else '94190a2d-d1c7-46c6-8144-da38141aa027'
+        email_template = 'c82b8ffd-f67c-4019-a724-d57ab559f08e' if early_years_register \
+                         else '94190a2d-d1c7-46c6-8144-da38141aa027'
         view_template = 'payment-confirmation-health-dbs.html'
 
     elif ((capita and certificate_information_boolean) or on_update) and not lived_abroad:
@@ -110,7 +115,8 @@ def get_template(crc, app_id, application, cost, cr_type):
         else:
             log.debug('Attempting send of email "Confirmation - DBS only"')
 
-        email_template = '02c01f75-1f9d-428f-a862-4effac03ebd3' if early_years_register else '294f2710-c507-4d30-ae64-b5451c59a45c'
+        email_template = '02c01f75-1f9d-428f-a862-4effac03ebd3' if early_years_register \
+                         else '294f2710-c507-4d30-ae64-b5451c59a45c'
         view_template = 'payment-confirmation-dbs-only.html'
 
     elif (capita and not certificate_information_boolean) and not lived_abroad:
@@ -119,7 +125,8 @@ def get_template(crc, app_id, application, cost, cr_type):
         else:
             log.debug('Attempting send of email "Confirmation - no docs to send"')
 
-        email_template = '8ca4eb7c-f4c9-417a-85e6-f4c10672f41a' if early_years_register else '75325ea2-c9b4-408c-9d89-c16ebbd7bd32'
+        email_template = '8ca4eb7c-f4c9-417a-85e6-f4c10672f41a' if early_years_register \
+                         else '75325ea2-c9b4-408c-9d89-c16ebbd7bd32'
         view_template = 'payment-confirmation-no-documents.html'
 
     else:
